Remove support for Chef 10 / 11 from chef-zero / chef-solo
The chef-zero provisioner provisioner had some crazy hacks in place to support Chef 11 < 11.8 by setting up chef-zero on its own. This removes that
The chef-solo provisioner had checks for Chef < 11 which have also been removed

It seems fair to assume that everyone is on Chef 12 in order to use modern Test Kitchen at this point and this removes a bunch of logic checks that had to run each  time Kitchen was used.

Signed-off-by: Tim Smith <tsmith@chef.io>4