Chef provisioners should prompt for license acceptance
The provisioner itself is not protected under the Chef Software EULA. But the product it installs on provisioned machines (Chef Client, ChefDK, etc.) is. So we perform a workstation-side early check to ensure the current user has accepted the license. The acceptance flow is the same as it is for the clients. Acceptance is automatically passed to provisioned instances.

To accept the license the user can:
* Set the CHEF_LICENSE environment variable
* Set the 'chef_license' attribute in the kitchen.yml under the provisioner
* If these are not set and the license has not been previously accepted an interactive prompt will occur asking the user to accept.

Possible values for acceptance are "accept" or "accept-no-persist".

License acceptance is stored on the workstation (unless the user sets "accept-no-persist"). Once stored the license acceptance does not need to be accepted as a flag.

Signed-off-by: tyler-ball <tball@chef.io>