Let license errors raise without trying to cleanup sandbox
Before this change if the license acceptance failed and raised an
exception the 'cleanup_sandbox' was called in an ensure. But the
sandbox would not have been created yet so it would fail with an error
like:

Failed to complete #converge action: [Sandbox directory has not yet been
created. Please run Kitchen::Provisioner::ChefGithub#create_sandox
before trying to access the path.]

Signed-off-by: tyler-ball <tball@chef.io>