:sparkles: Add instance name to download path
If the same download path is used for all instances, it may prove
problematic, especially if it is configured to download specific files;
the file would be overwritten.

Adding the instance name to the path allows the feature to be used by
multiple instances in succession without potentially overwriting files.