Move sudo from install_command down into lower methods
This prevents an issue in which sudo is only applied to the first command
in install_from_file(), which is now a mkdir. In that case, the directory
would be created as root, but the install script and error log would be
written as a regular user, resulting in a permission denied when trying
to write to a file named "stderr".

Running sudo on every command in install_from_file is an expansion of
privilege which we wish to avoid, and also makes the writing of a large
string to a file more complex than needed.

Signed-off-by: Clinton Wolfe <clintoncwolfe@gmail.com>