Support using the chef-cli gem to depsolve policyfiles
Previously you could install the chefdk gem to get the chef command line
and that could be used to depsolve policyfiles. That gem is now
deprecated in favor of the chef-cli gem, which ships a chef-cli cli.
This was done when we moved to a Go based chef CLI in Workstation that
just calls various sub-commands. With this change you can use
Policyfiles with a Test Kitchen install using Ruby gems. Right now it
will fail to find the Chef CLI

Signed-off-by: Tim Smith <tsmith@chef.io>