Use xfreerdp instead of rdesktop on Linux with WinRM
This works around the following problem with rdesktop:
```
ERROR: CredSSP: Initialize failed, do you have correct kerberos tgt initialized ?
Failed to connect, CredSSP required by server.
```

It seems to be related to this issue [1]. As a workaround, xfreerdp works out of
the box and seems to be a better option.

[1] https://github.com/rdesktop/rdesktop/issues/28

Signed-off-by: Lance Albertson <lance@osuosl.org>