Remove reference to duplicate catalog_id
catalog_id is mentioned in 2 places. It is confusing to find it as "catalog_id" and further down  the page as "catalog_id, catalog_name".   It is better to remove the first mention and leave the second as users should be aware of the catalog_id and the catalog_name as they are often used together.     

Signed-off-by: sc0ttruss <701447+sc0ttruss@users.noreply.github.com>